+++ driver/web/adapter.go
@@ -23,10 +23,9 @@ import (
 	"wellness/driver/web/rest"
 	"wellness/utils"
 
-	"github.com/rokwire/core-auth-library-go/v2/authservice"
-
 	"github.com/gorilla/mux"
-	"github.com/rokwire/core-auth-library-go/v2/tokenauth"
+	"github.com/rokwire/rokwire-building-block-sdk-go/services/core/auth"
+	"github.com/rokwire/rokwire-building-block-sdk-go/services/core/auth/tokenauth"
 	httpSwagger "github.com/swaggo/http-swagger"
 )
 
@@ -171,7 +170,7 @@ func (we Adapter) internalAuthWrapFunc(handler internalAuthFunc) http.HandlerFun
 }
 
 // NewWebAdapter creates new WebAdapter instance
-func NewWebAdapter(host string, port string, app *core.Application, config model.Config, serviceRegManager *authservice.ServiceRegManager) Adapter {
+func NewWebAdapter(host string, port string, app *core.Application, config model.Config, serviceRegManager *auth.ServiceRegManager) Adapter {
 	auth := NewAuth(app, config, serviceRegManager)
 
 	apisHandler := rest.NewApisHandler(app)